Add files via upload
Design a program that display items in a cafeteria. Allow the user to add items that he is purchasing, compute the total of items then add service taxes and tips afterwards.
Tip should be added after the tax is added up. Allow user to tip more if he wants.

Tax: 6.75%
Tips 10%

Display the items, charge, tax amount, tip amount and total bill on the screen.